move error component
